refactor: hooks decoupled from components for better maintenance
