+++ frontend/src/components/productForm/ProductForm.tsx
@@ -1,21 +1,7 @@
-import { useContext, useEffect, useState } from "react"
-import inventoryService from "../../services/inventory.service"
-import { InventoryItem, Product } from "../../types/inventory";
-import { StockStatus } from "../../utils/inventory.utils";
-import { InventoryContext } from "../../context/InventoryContext";
-import { PlusIcon } from "../../utils/icons";
+import { ProductFormProps } from "../../types/inventory"
+import { PlusIcon } from "../../utils/icons"
+import { useProductForm } from "../../hooks/useProductForm"
 
-/**
- * Props for the ProductForm component.
- *
- * @typedef {Object} ProductFormProps
- * @property {InventoryItem | null} [productToEdit] - Optional product to edit; if null, the form is for creating a new product.
- * @property {() => void} onClose - Function to call when closing the form.
- */
-interface ProductFormProps {
-  productToEdit?: InventoryItem | null; // Optional prop for editing
-  onClose: () => void; // Function to close the form
-}
 
 /**
  * ProductForm Component
@@ -43,97 +29,24 @@ const ProductForm = ({ productToEdit, onClose }: ProductFormProps) => {
   tomorrow.setDate(tomorrow.getDate() + 1)
   const minDate = tomorrow.toISOString().split('T')[0]
 
-  const validCharacters = /^[A-Za-z0-9 ]*$/
+  const productFormHook = useProductForm({ productToEdit, onClose })
 
-  const context = useContext(InventoryContext)
-  if (!context) {
-    return null
-  }
-  const { shouldUpdateTable, setShouldUpdateTable } = context.triggerTableUpdateType
+  const {
+    selectedName,
+    errorName,
+    categories,
+    isVisible,
+    allProductsAreValid,
 
-  const [selectedName, setSelectedName] = useState(productToEdit?.product.name || "")
-  const [errorName, setErrorName] = useState(false)
-  const [categories, setCategories] = useState<string[]>([]);
-  const [selectedCategory, setSelectedCategory] = useState(productToEdit?.product.category || "")
-  const [selectedStock, setSelectedStock] = useState(productToEdit?.quantity || 10)
-  const [selectedUnitPrice, setSelectedUnitPrice] = useState(productToEdit?.product.price || 0)
-  const [selectedDate, setSelectedDate] = useState<Date | null>(productToEdit?.product.expiryDate ? new Date(productToEdit.product.expiryDate) : null)
-  const [isVisible, setIsVisible] = useState(false)
-  const [allProductsAreValid, setAllProductsAreValid] = useState(true)
+    selectedCategory, setSelectedCategory,
+    selectedStock, setSelectedStock,
+    selectedUnitPrice, setSelectedUnitPrice,
+    selectedDate, setSelectedDate,
 
-  /**
-  * Fetches the list of product categories from the inventory service.
-  * The categories are updated whenever `shouldUpdateTable` changes.
-  */
-  useEffect(() => {
-    inventoryService.getCategories().then((response) => {
-      setCategories(response)
-    })
-  }, [shouldUpdateTable])
-  /**
-  * Handles form submission.
-  *
-  * Validates that required fields are provided and calls the appropriate
-  * inventory service method to either update an existing product or create a new one.
-  * Triggers a table update and closes the form if successful.
-  *
-  * @param {React.FormEvent} e - The form submission event.
-  */
-  const handleSumbit = (e: React.FormEvent) => {
-    e.preventDefault()
-    if (selectedName !== "" && selectedCategory !== "" && selectedStock !== 0 && selectedUnitPrice !== 0 && errorName === false) {
-      const product: Product = {
-        id: productToEdit?.id || 0,
-        name: selectedName,
-        category: selectedCategory,
-        price: selectedUnitPrice,
-        expiryDate: selectedDate ? new Date(selectedDate) : null,
-        dateCreated: new Date(),
-        dateUpdated: new Date()
-      }
-      const inventory: InventoryItem = {
-        id: productToEdit?.id || 0,
-        product: product,
-        quantity: selectedStock,
-        stockStatus: StockStatus.IN_STOCK
-      }
-
-      // Call the corresponding service method based on edit or creation mode
-      if (productToEdit) {
-        inventoryService.updateInventoryItem(inventory.id, inventory)
-        console.log("Item sent as: ", inventory)
-      } else {
-        inventoryService.createInventoryItem(inventory)
-      }
-      // Trigger table update and close the form
-      setShouldUpdateTable(prev => !prev)
-      onClose()
-    } else {
-      setAllProductsAreValid(false)
-    }
-  }
-
-  /**
-  * Toggles the visibility of the new category input.
-  */
-  const handleNewCategoryButton = () => {
-    setIsVisible(prev => !prev)
-  }
-  /**
-  * Handles changes in the product name input.
-  *
-  * Validates that the input contains only allowed characters.
-  *
-  * @param {React.ChangeEvent<HTMLInputElement>} e - The change event from the product name input.
-  */
-  const handleChangeName = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSelectedName(e.target.value)
-    if (!validCharacters.test(selectedName)) {
-      setErrorName(true)
-    } else {
-      setErrorName(false)
-    }
-  }
+    handleSumbit,
+    handleChangeName,
+    handleNewCategoryButton
+  } = productFormHook
 
   return <div>
     <p className="text-sm text-gray-600 font-light text-justify pb-2">Please provide the required information to register a new product.</p>
